refactor: OpenGraph 및 리팩토링
